Move Focus logic to Instrument/CameraInstrument
