+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "hidapi>=0.14.0",
     "colorama",
     "donjon_hyshlr",
+    "scipy",
     "pylmscontroller",
 ]
 requires-python = ">=3.9,<3.13"
